Define object DeltaTestImplicits
Define `object DeltaTestImplicits` (in unit test code base) and move the existing implicit class `DeltaTestUtils.OptimisticTxnTestHelper` there.

The change improves test code readability by making it obvious where the implicits come from: `import DeltaTestImplicits._` vs. importing classes with vague names or even silently inheriting them from a parent trait. It also provides a central location where we can define additional test-only implicits for Delta code -- several more are coming soon as part of an ongoing effort to clean up certain aspects of the Delta code base.

Test-only code; correct if unit tests continue to compile and pass.

GitOrigin-RevId: e32928c2b6274259b8e4938d4e3c0165c9d567ee
