+++ core/src/test/scala/org/apache/spark/sql/delta/DeltaRetentionSuite.scala
@@ -20,9 +20,9 @@ import java.io.File
 
 import scala.language.postfixOps
 
-import org.apache.spark.sql.delta.DeltaTestUtils.OptimisticTxnTestHelper
 import org.apache.spark.sql.delta.actions.{Action, AddFile, RemoveFile, SetTransaction}
 import org.apache.spark.sql.delta.test.DeltaSQLCommandTest
+import org.apache.spark.sql.delta.test.DeltaTestImplicits._
 import org.apache.spark.sql.delta.util.FileNames
 import org.apache.hadoop.fs.Path
